# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me/user": {
      "get": {
        "description": "Returns authenticated user profile including linked auth providers (Google, GitHub, etc).",
        "operationId": "ProfileController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Profile (candidate)"
        ]
      },
      "patch": {
        "description": "Updates first name, last name, and username.",
        "operationId": "ProfileController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation or username conflict error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user profile",
        "tags": [
          "Profile (candidate)"
        ]
      },
      "delete": {
        "description": "Marks the user account as suspended.",
        "operationId": "ProfileController_deactivateAccount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate account",
        "tags": [
          "Profile (candidate)"
        ]
      }
    },
    "/me/user/cooldown": {
      "get": {
        "description": "Returns authoritative cooldown timestamps for the authenticated user.",
        "operationId": "ProfileController_getCooldowns",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CooldownResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get analysis and sync cooldowns",
        "tags": [
          "Profile (candidate)"
        ]
      }
    },
    "/me/user/candidate": {
      "get": {
        "description": "Returns candidate-specific data including bio, career path, and dev profile.",
        "operationId": "ProfileController_getCandidateProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Candidate profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get candidate profile",
        "tags": [
          "Profile (candidate)"
        ]
      },
      "patch": {
        "description": "Updates bio and career path.",
        "operationId": "ProfileController_updateCandidateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update candidate profile",
        "tags": [
          "Profile (candidate)"
        ]
      }
    },
    "/me/user/github": {
      "get": {
        "description": "Returns whether GitHub is connected and sync status (without exposing tokens).",
        "operationId": "ProfileController_getConnectedGithub",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubConnectionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Candidate profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get GitHub connection status",
        "tags": [
          "Profile (candidate)"
        ]
      }
    },
    "/me/user/wallet": {
      "get": {
        "description": "Returns whether wallet is connected and sync status to which address.",
        "operationId": "ProfileController_getConnectedWallet",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Web3ConnectionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Candidate profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get wallet connection status",
        "tags": [
          "Profile (candidate)"
        ]
      }
    },
    "/me/user/avatar": {
      "post": {
        "operationId": "ProfileController_uploadAvatar",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Profile (candidate)"
        ]
      }
    },
    "/me/user/waitlist": {
      "post": {
        "description": "Reads email from the authenticated user record; no body required. Queues a confirmation email. Idempotent.",
        "operationId": "ProfileController_registerWaitlistAuth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join employer launch waitlist (authenticated)",
        "tags": [
          "Profile (candidate)"
        ]
      }
    },
    "/profile/public/{username}": {
      "get": {
        "operationId": "PublicProfileController_getPublicProfile",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Public Profiles"
        ]
      }
    },
    "/profile/public": {
      "get": {
        "operationId": "PublicProfileController_searchPublicProfiles",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Public Profiles"
        ]
      }
    },
    "/profile/waitlist": {
      "post": {
        "description": "Stores the email and sends a confirmation email. Idempotent — calling twice with the same email is a no-op.",
        "operationId": "PublicProfileController_registerWaitlist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWaitlistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Joined the waitlist"
          },
          "400": {
            "description": "Invalid email address"
          }
        },
        "summary": "Join employer launch waitlist (guest)",
        "tags": [
          "Public Profiles"
        ]
      }
    },
...
  },
  "info": {
    "title": "a16zero Employer API",
    "description": "Backend MVP for Employer platform features and Account Abstraction auth verification.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "internal-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-key",
        "description": "Internal API key"
      }
    },
    "schemas": {
...
      "AuthAccountDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "google"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-01T12:00:00.000Z"
          }
        },
        "required": [
          "provider",
          "createdAt"
        ]
      },
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string"
          },
          "accountStatus": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "mfaEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "authAccounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthAccountDto"
            }
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "firstName",
          "lastName",
          "name",
          "role",
          "accountStatus",
          "isEmailVerified",
          "mfaEnabled",
          "createdAt",
          "updatedAt",
          "authAccounts"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "CooldownItemDto": {
        "type": "object",
        "properties": {
          "cooldownUntil": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "cooldownUntil"
        ]
      },
      "CooldownResponseDto": {
        "type": "object",
        "properties": {
          "github": {
            "$ref": "#/components/schemas/CooldownItemDto"
          },
          "wallet": {
            "$ref": "#/components/schemas/CooldownItemDto"
          },
          "generate": {
            "$ref": "#/components/schemas/CooldownItemDto"
          }
        },
        "required": [
          "github",
          "wallet",
          "generate"
        ]
      },
      "SimpleMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "CandidateProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "careerPath": {
            "type": "number"
          },
          "scorecard": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "vouches": {
            "type": "object",
            "nullable": true
          },
          "devProfile": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "bio",
          "location",
          "website",
          "careerPath",
          "scorecard",
          "createdAt",
          "vouches",
          "devProfile"
        ]
      },
      "UpdateCandidateDto": {
        "type": "object",
        "properties": {}
      },
      "GithubConnectionResponseDto": {
        "type": "object",
        "properties": {
          "connected": {
            "type": "boolean"
          },
          "github": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "connected",
          "github"
        ]
      },
      "Web3ConnectionResponseDto": {
        "type": "object",
        "properties": {
          "connected": {
            "type": "boolean"
          },
          "web3": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "connected",
          "web3"
        ]
      },
      "RegisterWaitlistDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "candidate@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  }
}